Move common modules to src/maxtext/common
