+++ src/MaxText/decode.py
@@ -26,9 +26,9 @@ from jetstream.engine import engine_api
 from MaxText import max_utils
 from MaxText import maxengine
 from MaxText import pyconfig
-from MaxText import profiler
 from MaxText import multimodal_utils
 from MaxText.multimodal import preprocessor
+from maxtext.common import profiler
 # Placeholder: internal
 
 # Number of text sequences to process in a single batch.
