+++ src/MaxText/inference_microbenchmark.py
@@ -26,9 +26,9 @@ from MaxText import max_utils
 from MaxText import maxengine
 from MaxText import maxtext_utils
 from MaxText import prefill_packing
-from MaxText import profiler
 from MaxText import pyconfig
 from MaxText.utils import gcs_utils
+from maxtext.common import profiler
 
 import warnings
