+++ src/MaxText/maxtext_utils.py
@@ -35,7 +35,6 @@ import optax
 import orbax.checkpoint.experimental.emergency.checkpoint_manager as emergency_checkpoint_manager
 import orbax.checkpoint.experimental.emergency.replicator_checkpoint_manager as emergency_replicator_checkpoint_manager
 
-from MaxText import checkpointing
 from MaxText import max_logging
 from MaxText import max_utils
 from MaxText import multimodal_utils
@@ -43,6 +42,7 @@ from MaxText import sharding
 from MaxText.configs import types
 from MaxText.common_types import DecoderBlockType, MODEL_MODE_PREFILL, MODEL_MODE_AUTOREGRESSIVE
 from MaxText.inference.page_manager import PageState
+from maxtext.common import checkpointing
 
 OVERWRITE_WITH_GRADIENT = "_overwrite_with_gradient"
