+++ src/MaxText/utils/ckpt_conversion/to_maxtext.py
@@ -74,7 +74,6 @@ from safetensors import safe_open
 import absl
 
 from orbax.checkpoint import type_handlers
-from MaxText import checkpointing
 from MaxText import max_logging
 from MaxText import max_utils
 from MaxText import maxtext_utils
@@ -82,9 +81,9 @@ from MaxText import pyconfig
 from MaxText.common_types import MODEL_MODE_TRAIN
 from MaxText.inference_utils import str2bool
 from MaxText.layers import models, quantizations
-from MaxText.checkpointing import save_checkpoint
 from MaxText.utils.ckpt_conversion.utils.param_mapping import HOOK_FNS, PARAM_MAPPING
 from MaxText.utils.ckpt_conversion.utils.utils import apply_hook_fns, HF_IDS, print_ram_usage, get_hf_model, validate_and_filter_param_map_keys
+from maxtext.common import checkpointing
 
 jax.config.update("jax_platform_name", "cpu")
 
@@ -729,7 +728,7 @@ def main(args: Sequence[str], test_args: Sequence[str]) -> None:
     else:
       max_logging.log("Starting checkpoint save...")
 
-    if save_checkpoint(checkpoint_manager, 0, final_save_state):
+    if checkpointing.save_checkpoint(checkpoint_manager, 0, final_save_state):
       max_logging.log("saved a checkpoint at step 0")
 
     # Upon preemption, exit when and only when all ongoing saves are complete.
