+++ src/MaxText/utils/ckpt_scripts/convert_gemma2_chkpt.py
@@ -32,7 +32,7 @@ from flax.training import train_state
 import orbax
 
 from MaxText import max_logging
-from MaxText import checkpointing
+from maxtext.common import checkpointing
 
 jax.config.update("jax_platform_name", "cpu")
