+++ src/MaxText/utils/ckpt_scripts/convert_gpt3_ckpt_from_paxml.py
@@ -49,7 +49,6 @@ from jax.sharding import Mesh
 
 import tensorstore as ts
 
-from MaxText import checkpointing
 from MaxText import max_logging
 from MaxText import maxtext_utils
 from MaxText import max_utils
@@ -59,6 +58,7 @@ from MaxText.common_types import MODEL_MODE_TRAIN
 from MaxText.globals import MAXTEXT_PKG_DIR
 from MaxText.layers import quantizations
 from MaxText.layers.models import transformer_as_linen
+from maxtext.common import checkpointing
 
 
 def fmt_size(num_bytes: int) -> str:
