+++ src/MaxText/utils/lora_utils.py
@@ -22,11 +22,11 @@ import jax.numpy as jnp
 from flax.training import train_state
 from flax.linen import partitioning as nn_partitioning
 
-from MaxText import checkpointing
 from MaxText import max_utils
 from MaxText import maxtext_utils
 from MaxText import max_logging
 from MaxText.utils import gcs_utils
+from maxtext.common import checkpointing
 
 
 def apply_lora_on_base_params(base_params, lora_params, lora_scale_factor=1.0):
