+++ src/maxtext/common/metric_logger.py
@@ -32,9 +32,8 @@ from MaxText import max_utils
 from MaxText import maxtext_utils
 from MaxText.managed_mldiagnostics import ManagedMLDiagnostics
 from MaxText.utils import gcs_utils
-from MaxText.gcp_workload_monitor import GCPWorkloadMonitor
 from MaxText.globals import EPS
-
+from maxtext.common.gcp_workload_monitor import GCPWorkloadMonitor
 from collections import defaultdict
 
 # Mapping MaxText metrics to managed profiler metrics
