+++ tests/unit/data_loader_test.py
@@ -23,12 +23,12 @@ import jax
 from unittest.mock import MagicMock
 from jax.sharding import Mesh
 
-from MaxText.data_loader import DataLoader, RampUpDataLoader
 from MaxText.rampup_batch import RampupBatchManager
 from MaxText.maxtext_utils import create_device_mesh
 from MaxText import exceptions
 from MaxText import pyconfig
 from MaxText.globals import MAXTEXT_PKG_DIR
+from maxtext.common.data_loader import DataLoader, RampUpDataLoader
 
 
 class DataLoaderTest(unittest.TestCase):
